requirements.txt
Pillow==9.4.0
pandas==2.0.1
torch==1.13.1
numpy==1.24.1
scipy==1.9.3
statsmodels==0.14.0
POT==0.9.0
tqdm==4.60.0
torchvision==0.14.1
scikit-learn==1.1.3
matplotlib==3.6.2
seaborn==0.12.2
pytorch-lightning==1.5.6
lightning-bolts==0.3.2
accelerate==0.15.0
hydra-core==1.0.6
